backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
faker==22.6.0

# NLP & AI
spacy==3.7.5
scikit-learn==1.1.2
scispacy==0.5.4
negspacy==1.0.4
transformers==4.37.2
torch==2.2.0
sentence-transformers==2.3.1

# Vector DB
chromadb==0.4.22

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.8
smart-open==6.4.0

# OCR (optional - system will work without these)
Pillow==10.2.0
pytesseract==0.3.10

# Data Processing
pandas==2.2.0
numpy==1.26.3
packaging==24.2


# HTTP Client
httpx>=0.27,<0.29
requests==2.31.0

# LLM Integration
langchain==0.2.1
langchain-community==0.2.1
langchain-ollama==0.1.1


# Address Parsing
usaddress==0.5.10

# Validation
jsonschema==4.20.0

# Utilities
python-dateutil==2.8.2
neo4j==5.12.0
biopython==1.81
rdkit==2023.9.4
pubchempy==1.0.4

pdf2image==1.17.0
